Explicitly separate region deletion code
The garbage collection of read/write regions must be separate from
read-only regions:

- read/write regions are garbage collected by either being deleted
  during a volume soft delete, or by appearing later during the "find
  deleted volume regions" section of the volume delete saga

- read-only regions are garbage collected only in the volume soft delete
  code, when there are no more references to them

`find_deleted_volume_regions` was changed to only operate on read/write
regions, and no longer returns the optional RegionSnapshot object: that
check was moved from the volume delete saga into the function, as it
didn't make sense that it was separated.

This commit also adds checks to validate that invariants related to
volumes are not violated during tests. One invalid test was deleted
(regions will never be deleted when they're in use!)

In order to properly test the separate region deletion routines, the
first part of the fixes for dealing with deleted volumes during region
snapshot replacement were brought in from that branch: these are the
changes to region_snapshot_replacement_step.rs and
region_snapshot_replacement_start.rs.