Add clickhouse policy to the db (#6862)
This enables us to enable and disable both clickhouse clusters and
single-node clickhouse zones.

We anticipate few changes to this policy and so we maintain a complete
history indexed by version. We only allow updates via a single version
increment, and we guarantee consistency by ensuring the previous version
is the latest version in the database while updating.