Check connection health on recycle
Adds in a call to ping the server in the `Connector::on_recycle()`
method, if we haven't already validated the connection by cancelling a
query. This means that qorb learns basically right away that a
connection is broken, and tries to reestablish it more quickly. This
obviates some of the mucking about with timeouts in earlier commits.