Add ClickHouse native connection pool (#6889)
- Implement a connector and connection pool for talking to ClickHouse
over the native TCP protocol
- Add the native TCP address or a `BoxedResolver` for resolving it to
all code that connects to ClickHouse. This is a pretty noisy change,
since we now need two addresses / objects everywhere. We'll take it back
down to one when we completely make the switch to the native protocol
and remove the HTTP interface.
- Remove the feature flag gating the native code, keeping that just for
the prototype SQL client shell.
- NFC formatting change to bring the `oximeter_db::native` import style
in line with the rest of the crate.