[reconfigurator-cli] disallow loading state into non-empty systems (#6891)
Currently, the `cmd_load` command doesn't require an empty system and
lets you effectively merge two states. However, one of the things it
does is overwrite the internal and external DNS maps, which means that
existing blueprints can end up pointing to DNS generations that don't
have any correspondence.

The easiest way to address this issue is to change the `load` function
to require an empty system, similar to `load-example`. Make that change
here, and add some tests for it.

The load function also becomes somewhat simpler, since a bunch of checks
that currently exist can go away.