Remove the LazyTimeseriesClient in Nexus (#6925)
The `LazyTimeseriesClient` is a vestigial type we've used to connect to
ClickHouse on-demand. This is used in Nexus, which may be created and
start handling requests that require talking to ClickHouse before the
server may be up or reachable.

That's obviated by the `qorb` connection pooling we've added in recent
commits. But since it's still around, that means we create connection
pools _each time_ Nexus handles a timeseries-related request. That leads
to exorbitant numbers of connections, almost all of which timeout, but
which causes ClickHouse to be unable to handle new connections. This
closes #6923.