[omdb] Allow selecting instances by state (#6930)
Presently, `omdb db instances` has a `--running` flag, which causes it
to only show instances which are in the "Running" external API state. In
many cases it is desirable to select only instances which are in other
states. This branch adds a `--state` flag, which accepts any number of
DB instance states to select. Multiple `--state` arguments can be
provided. The previous `--running` flag is still accepted for
backwards-compatibility purposes, since it seemed like there wasn't
really any reason to get rid of it.

Fixes #6927