Fix: use correct primary NIC for External IP add/delete/refresh (#6890)
`InstanceRunner` has been using the wrong primary NIC for
Ephemeral/Floating IP addition/removal for some time, assuming that the
Slot 0 NIC is always the primary. Now that External IP state is
regularly re-asserted to keep Internet Gateway tags up-to-date, this led
to OPTE ports reporting the use of external IPs on both the primary and
slot0 nic, when multiple NICs are present and external IP state is
changed on the instance.

This PR makes the correct NIC selection for all three operations.