[dns-server] only return answers to the incoming question (#6308)
As noted in https://github.com/oxidecomputer/omicron/issues/4051,
queries to the internal DNS server would get any records we have for a
name in response, rather than only records matching the query incoming
query type. That behavior is confusing, but worse, wrong.

Subtly, this is not actually a misbehavior I believe we can observe
through `trust_dns_resolver`: the resolver from that crate includes its
own `CachingClient`. As a side effect of upstream answers going through
that caching client, incorrect Answers records are cached and only
correct answers actually make it out to us as consumers of
`trust_dns_resolver`.

But as is plenty clear in #4051, `dig` and other DNS clients can get
incoherent answers!

Simple enough to fix: only return answers that are answers to the
question we were asked.