[api] Improve instance `auto_restart_policy` docs
Currently, the documentation for the `auto_restart_policy` field in the
`params::InstanceCreate` and `params::InstanceUpdate` types doesn't
really explain the meaning of a `null` auto-restart policy. The comment
on the `AutoRestartStatus` type in the instance *view* API does explain
that `null` means "use whatever the default currently is", but this
isn't obvious on the APIs where a user would actually *set* the policy
to `null`.

This was an oversight on my part, and this commit corrects it. Thanks
@davidcrespo for noticing!

Fixes #6957