[omdb] Make `db instance info` less cryptic (#6948)
When displaying an instance's reincarnation status in `omdb db instance
info`, I was a bit too cutesy with the Buddhism references ---
currently, it prints that an instance is "bound to saṃsāra" or has
"attained nirvāṇa", which is not the clearest description of the
instance's reincarnation policy. This commit changes the `omdb db
instance info` output a bit to try and make the instance's reincarnation
policy clearer.