[nexus] Manage datasets via reconfigurator (#6229)
This PR makes the blueprint aware of datasets, and manages them as a
part of planning and execution.

As a part of the planning phase: the reconfigurator ensures that all
in-service disks have debug / zone root datasets, and that all zones
have the corresponding datasets they need to function. Additionally,
unused datasets are marked as "expunged" until they are gone from the
database, at which point they're fully deleted from the blueprint.

As a part of the execution phase: the reconfigurator ensures that
database records exist for these datasets, and requests them from
appropriate sled agents. Additionally, "expunged" datasets are removed
from the database explicitly.

Fixes #6043 
Fixes #6107

---------

Co-authored-by: Rain <rain@oxide.computer>