[nix flake] Update lockfile
As of #6949, the `rust-toolchain.toml` file for Omicron uses Rust
1.82.0. This Rust release was not in nixpkgs at the last time the Nix
flake's lockfile was updated, so in order for Nix users to get the
correct toolchain, the flake must be updated. This commit includes the
results of running `nix flake update`, which updates the lockfile to a
Nixpkgs revision that includes the desired toolchain.