Delay starting zone-network-service until in.ndpd is online
There is a race condition between being able to configure an
addrconf address on an interface, and the NDP daemon starting up.
If they start in parallel, there is a chance that configuring
network interfaces will fail resulting in a broken zone. There
is a fix to implement better handling of the conflict in stlouis
but we should also just wait for the NDP service to come up before
attempting to configure interfaces.

Fixes #6947 (along with stlouis#650)