[omdb] Display (some) name columns last in tables
Presently, OMDB lists the names of instances and disks towards the
left-hand of tabular output. This is probably because instance/disk
names are relatively _important_ to the user, so it makes sense to put
them first. However, user-defined string names also wildly vary in
length, in contrast to many other columns in OMDB output, such as UUIDs,
timestamps, etc, which are broadly fixed width. Putting names "early"
(from a left-to-right perspective) means that when there are a couple
names in the list that are much longer than the majority of names, most
lines have a lot of whitespace to pad the short names to the same length
as the long ones. This wastes horizontal characters and makes the output
more likely to wrap on reasonably sized terminals. Putting the name
column last makes it likelier that only those names that are very long
will line-wrap, and most lines will not wrap, which makes tabular output
much more intelligible.