convert Nexus to new sled agent VMM API
This ends up touching a substantial number of files because of the way
Nexus accessed Crucible client types:

- Nexus imported Crucible types from the generated sled agent client
- Sled agent's code included generated Crucible types from the Propolis
  client
- The Propolis client generates Crucible types because the Propolis
  server API used a DiskRequest type that referred to the Crucible
  client library's VolumeConstructionRequest and CrucibleOpts types

The Propolis server API no longer has a DiskRequest type, so generated
Propolis clients no longer include the Crucible types, and neither do
the generated sled agent clients.

Deriving Crucible types from the Propolis client is useful because it
ensures that Nexus's conception of a VolumeConstructionRequest is the
same as Propolis's. To maintain this property, the Propolis client
depends on the crucible-client-types crate and re-exports types from it;
the sled agent client then re-exports from the Propolis client
library.

The catch is that Progenitor-generated versions of these types use
Strings for socket addresses, while the native types use native
SocketAddrs. This requires a fair bit of munging in the tests to remove
some conversions between strings and socket addresses.