rework boot disk update with better filtering
"better" here meaning "if we don't need to update the database, don't
update the database". if we would change the boot disk to the same
value, or we would change it from null to null, this change means we
will instead not even select the row for an update. since we still
`check_if_exists` we'll still fetch an Instance for the operation if
there's no update performed. so, if the database can detect that the
update would have been a no-op and elide it, this change is not
particularly useful. if the database would have to commit some state as
a result of an ineffectual `set`, then this is helpful!

more importantly, this is moderately more legible than smearing the
update logic and consistency checking between the Diesel query and
handling in the `NotUpdatedButExists` arm.

at the same time, Eliza rightfully pointed out that there are now (at
least) two copies of a list of states that are both roughly "there is no
VMM for this instance, so we can do things to its' database record".
extract that out to a const on `InstanceState` and name it
appropriately.

finally, the instance size change logic incorrectly rejected size
changes that only changed one of `(memory, cpus)`. fix that to allow
changing one or the other independently and add tests to catch a
future dingus (me) who might regress that somehow.