move back to Rust 1.81 (#7020)
Rust 1.82 has a rather serious compile time regression in Diesel-generated
code: https://github.com/rust-lang/rust/issues/132064. This makes Omicron
builds up to twice as slow.

We used to be on 1.80 before moving to 1.82, but 1.81 has some nice
things like `#[expect]`. So move to 1.81.

We expect the regression to be fixed in Rust 1.84.

Thanks to @jmpesp for tracking this down!