Insert samples using the native client (#7013)
- Remove unneeded and error-prone `n_columns` and `n_rows` fields from
the `Block` type. These are now methods that delegate to the actual
column arrays, which are the source of truth anyway.
- Add methods for extracting fields from a sample as a set of data
blocks, one per field table.
- Add methods to extract measurements from a sample as a block destined
for one table. Take care to correctly construct missing samples,
especially for histograms. Also implement the `FromBlock` trait for
samples, to ensure we can extract the raw data as well. This is only
used in tests, in this commit.
- Insert fields and measurements from a sample as a data block, using
the native interface.
- Fix serde of UUIDs in native format, which doesn't match the
documentation at least for our current version of ClickHouse.
- Remove code serializing fields and measurements into JSON.
- Closes #6884