Add clickhouse-cluster-config to omdb blueprint output (#6968)
Print clickhouse cluster config related tables for `blueprint show` and
`blueprint diff` omdb commands.

A bunch of the complexity and duplication here arises from the fact
that  we are diffing not only between blueprints that have identical
structures, but between blueprints and collections that have
drastically different contents. This is useful, but we probably should
consider separating the two types of diffs and reworking all the
blueprint diff logic to use some sort of semantic diff between types
such as https://github.com/distil/diffus as recommended by @sunshowers.

In order to make the `cluster_secret` UUID generation deterministic for
tests I had use an rng seed in the `BlueprintBuilder`. This required
moving
creation of the initial `ClickhouseClusterConfig` and it's wrapping 
`ClickhouseAllocator` from `BlueprintBuilder::new_based_on` to 
`BlueprintBuilder::build`. 

Fixes #6941