simplify inputs to BuilderExternalNetworking
Removes three tests that aren't particularly useful: they're validating
that the builder rejects bogus PlanningInputs, but since those tests
were written we've updating PlanningInput itself to reject those inputs.
With the rearranging done in this PR it's not possible to construct a
PlanningInput that causes these tests to fail in the same way.