[nexus] Move existing OxQL endpoints under `/v1/system/*` (#7047)
Discussed this with @bnaecker a couple of weeks ago. Was going to do
this in #6873 but I decided to split it out because it's just noise. The
idea here is that (for now) rather than getting elaborate with
permissions on the OxQL endpoint a la #5298, we just keep the existing
endpoint fleet-viewer scoped and move it to
`/v1/system/timeseries/query` with all the other operator endpoints.
#6873 goes on top of this one and adds a project-scoped endpoint at
`/v1/timeseries/query?project=my-proj`.