Refactor clickhouse admin integration tests (#7017)
## Overview

This commit changes the way the clickhouse clusters are spun up for
clickhouse-admin integration tests. Instead of multiple clusters being
spun up for every test, now we start up a single cluster via nextest's
"setup scripts" functionality.
This way, we no longer need to manually hard code so many ports per
test.

## Caveats

Sadly, we still have to hardcode some ports, but at least this way it's
only a set of ports, and we won't need to add more every time we want to
test something new.

In order to get the teardown function working, I had to limit the tests
to run on a single thread. They're a bit slower than before, but nothing
too terrible. Setting up the cluster, running the tests and tearing down
takes ~7s. The time to run the tests themselves is ~3s