[reconfigurator] introduce a simulator, make CLI use it (#7022)
For randomized testing, it's useful to have a higher-level notion of a
succession of system states, which can be inspected, rewound, and other
branches explored as desired. Introduce this via a new `Simulator`
struct.

The individual states are roughly identical to the existing
`ReconfiguratorSim` in the CLI, and are a pretty standard tree
structure. For
now I've chosen not to use a Merkle tree, just a UUID-based tree, but if
there's a good use case for it we may choose to make it a Merkle tree in
the
future.

I've added a few basic features like storing the log of changes (though it's
not complete... I think I'll want to make `SystemDescription` follow the same
structure with read-only and mutable versions.)

As a proof of concept, this PR also rebuilds the current reconfigurator-cli
script on top of the simulator. I've not made any functional changes to the
CLI, but hope to add features like looking at the tree of states and undoing
them, and with randomized testing to be able to load up the tree
of states.

The simulator uses a bunch of structural sharing via `Arc`. I was wondering
whether I could use the `im` crate, but it doesn't seem to have an `IndexMap`
equivalent. The `Arc`s should make cloning cheap enough that I don't think it's
an issue in any case. (There's a comment in the PR about why I haven't used an
object-store-like style instead -- it's just easier to use `Arc` for
now.)

There's a bunch of work to be done before this can be landed, but I wanted to
put it up for folks to have an early look.

I've left a few TODOs in for future work. I don't think any of them block this
PR.