Remove `retry_until_known_result()` (#6868)
This builds on #6866. After its changes, there was only caller of
`retry_until_known_result` left; this PR removes it. We keep the retry
loop, but instead of retrying forever, we bail out if the sled we're
trying to reach is gone, as determined by "is it no longer in-service",
which in practice means it's been expunged.