Single node clickhouse init (#6903)
Depends on #6894, replaces #6878, fixes #6826 on the server side.

Adds a new Dropshot server, `clickhouse-admin-single`, analogous to
`clickhouse-admin-keeper` and `clickhouse-admin-server` (which were
split off from `clickhouse-admin` in #6837). Its sole purpose is to
initialize the single-node ClickHouse database with the current Oximeter
schema. We use a single in-memory lock (a `tokio::sync::Mutex`) to
serialize initialization requests. Multi-node ClickHouse clusters will
need something analogous as a follow-up.

---------

Co-authored-by: Andrew J. Stone <andrew@oxidecomputer.com>
Co-authored-by: Karen Cárcamo <karencfv@users.noreply.github.com>