[sled-storage] Automatically clean up test zpools, storage (#7111)
`StorageManagerTestHarness` is a utility which allows sled agent tests
to use real zpools and datasets atop temporary file-backed storage. This
provides a decent test fidelity for storage-based parts of the sled
agent.

This harness previously took a liberal attitude towards zpool creation,
but a conservative perspective on their deletion: rather than
automatically destroying zpools on test failure, it would print commands
for "how to delete these zpools yourself". This tendency was mostly
borne out of fear: deleting zpools is a dangerous operation which should
be done with caution.

When working with the `StorageManagerTestHarness`, especially while
iterating on tests with many expected failures, I've decided this is an
Enormous Pain In The Butt. This PR changes the aforementioned tendency,
and just attempts to delete all provisioned zpools we created in our
tests. If we fail to do so, we'll *then* log an error message and let
the user take responsibility for manual cleanup.