Remove HTTP ports from the `oximeter_db::Client` (#7118)
- Remove the HTTP address argument from the client constructor
- Remove the HTTP resolver from the client constructor
- Remove the HTTP pool from the client
- Rename `native_{address,port}` to just `{address,port}` in
configuration and CLI flags
- Closes #7094
- NOTE: We're still populating the DNS records for the HTTP address, and
relying on the implicit addition of the native TCP address via
`DnsBuilder::service_zone_clickhouse()`. We're also specifying the HTTP
address in the SMF properties of the ClickHouse zone. We don't strictly
need these anymore, but it's a bit more complicated to remove them,
since some deployed systems rely on those records. That can be done in a
follow-up, in which we'll switch from specifying the HTTP to the native
addresses.